[fix] SID: review fixes + fail-fast validation; fix integration test CPU pin
- sid_integration_test: force CPU with CUDA_VISIBLE_DEVICES="-1" not ""
  (empty is treated inconsistently across CUDA runtimes; the GPU CI runner
  didn't hide devices, tripping the CPU-only guard in the train_eval child).
- BaseSidModel: validate codebook entries >=1 and input_dim >=1 at construction;
  guard feature width in _extract_feature (a (B,1) tensor would otherwise
  broadcast into a degenerate rank-1 codebook). assert -> raise.
- residual_kmeans_quantizer / kmeans_quantize: assert -> raise for the
  data-corruption guards (N>=max_k, load_centroids_ shape, CPU/shape contract)
  so they survive python -O.
- RelativeL1: float64 sum / long count to avoid float32 rounding past 2**24.
- kmeans_quantize: drop the duplicate relative_l1/recon_diagnostics helpers;
  RelativeL1 (tzrec/metrics) is the single home of the formula. Per-layer
  offline-fit log now reports MSE only.
- sid_rqkmeans: TODO documenting the periodic-checkpointing-disabled contract
  (codebook can be dropped by save dedupe otherwise).
- sid_model.proto: drop stale "(global, across all ranks)" wording.
- mock config: set save_checkpoints_steps/epochs = 0 (the documented convention).

+++ tzrec/metrics/relative_l1.py
@@ -29,8 +29,17 @@ class RelativeL1(Metric):
     def __init__(self, epsilon: float = 1e-4, **kwargs) -> None:
         super().__init__(**kwargs)
         self.epsilon = epsilon
-        self.add_state("sum_rel", default=torch.tensor(0.0), dist_reduce_fx="sum")
-        self.add_state("count", default=torch.tensor(0.0), dist_reduce_fx="sum")
+        # float64 sum / long count: element-wise aggregation crosses 2**24 at
+        # only ~32K rows of a 512-dim embedding, past which float32 increments
+        # round (mirrors the float64 care in ``ReservoirSampler.add``).
+        self.add_state(
+            "sum_rel",
+            default=torch.tensor(0.0, dtype=torch.float64),
+            dist_reduce_fx="sum",
+        )
+        self.add_state(
+            "count", default=torch.tensor(0, dtype=torch.long), dist_reduce_fx="sum"
+        )
 
     def update(self, preds: torch.Tensor, target: torch.Tensor) -> None:
         """Accumulate the relative-L1 error between ``preds`` and ``target``.
@@ -42,7 +51,7 @@ class RelativeL1(Metric):
         rel = torch.abs(target - preds) / (
             torch.maximum(torch.abs(target), torch.abs(preds)) + self.epsilon
         )
-        self.sum_rel += rel.sum()
+        self.sum_rel += rel.sum().double()
         self.count += rel.numel()
 
     def compute(self) -> torch.Tensor:
